emo/requirements.txt
Django>=3.0,<4.0
requests
djangorestframework
librosa==0.8.1
tensorflow
Keras==2.3.1
numpy==1.16.0